feat(checkstyle): #1770 prohibit static nested classes outside tests
Neither Checkstyle nor PMD ship a rule for this, so it needed a
custom check. A non-test class may no longer nest a static class;
it must live in its own file, since a static nested class is a
self-contained abstraction. Non-static inner classes and static
nested classes inside test files (*Test, *IT, *ITCase) remain
allowed.

Enabling the check against Qulice's own sources required
extracting its own static nested helper classes into standalone
files.
