+++ src/main/resources/com/qulice/checkstyle/checks.xml
@@ -475,6 +475,7 @@
       <property name="excludeFileNamePattern" value=".*(Test|IT|ITCase)\.java"/>
     </module>
     <module name="com.qulice.checkstyle.ProhibitFieldsInTestClassesCheck"/>
+    <module name="com.qulice.checkstyle.ProhibitStaticNestedClassesCheck"/>
     <module name="com.qulice.checkstyle.ProtectedMethodInFinalClassCheck"/>
     <module name="com.qulice.checkstyle.QualifyInnerClassCheck"/>
     <module name="com.qulice.checkstyle.FinalSemicolonInTryWithResourcesCheck"/>
