+++ src/test/java/com/qulice/maven/MavenEnvironmentMocker.java
@@ -5,22 +5,15 @@
 package com.qulice.maven;
 
 import com.qulice.spi.Environment;
-import java.io.File;
 import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedList;
-import java.util.Map;
-import java.util.Properties;
 import org.apache.maven.monitor.logging.DefaultLog;
 import org.apache.maven.project.MavenProject;
 import org.codehaus.plexus.DefaultPlexusContainer;
 import org.codehaus.plexus.PlexusContainer;
 import org.codehaus.plexus.PlexusContainerException;
-import org.codehaus.plexus.context.Context;
-import org.codehaus.plexus.context.ContextException;
 import org.codehaus.plexus.logging.console.ConsoleLogger;
 import org.slf4j.impl.StaticLoggerBinder;
 
@@ -140,168 +133,11 @@ public final class MavenEnvironmentMocker {
         this.prj.inBasedir(this.ienv.basedir());
         return new MavenEnvironment.Wrap(
             this.ienv,
-            new MavenEnvironmentMocker.FakeMavenEnvironment(
+            new FakeMavenEnvironment(
                 this.prj.mock(),
-                new MavenEnvironmentMocker.FakeContext(this.container),
+                new FakeContext(this.container),
                 this.ass
             )
         );
-    }
-
-    /**
-     * FakeContext.
-     * A mock to a context.
-     * @since 0.24.1
-     */
-    private static final class FakeContext implements Context {
-
-        /**
-         * Container.
-         */
-        private final PlexusContainer container;
-
-        FakeContext(final PlexusContainer ctainer) {
-            this.container = ctainer;
-        }
-
-        @Override
-        public boolean contains(final Object obj) {
-            return true;
-        }
-
-        @Override
-        public void put(final Object obja, final Object objb) {
-            // Intentionally left blank
-        }
-
-        @Override
-        public Object get(final Object obj) throws ContextException {
-            return this.container;
-        }
-
-        @Override
-        public Map<Object, Object> getContextData() {
-            return Collections.emptyMap();
-        }
-    }
-
-    /**
-     * FakeMavenEnvironment.
-     * A mock to MavenEnvironment.
-     * @since 0.24.1
-     */
-    private static final class FakeMavenEnvironment implements MavenEnvironment {
-
-        /**
-         * Project.
-         */
-        private final MavenProject proj;
-
-        /**
-         * Context.
-         */
-        private final Context ctx;
-
-        /**
-         * Asserts.
-         */
-        private final Collection<String> assrts;
-
-        FakeMavenEnvironment(
-            final MavenProject prj,
-            final Context ctx,
-            final Collection<String> asserts
-        ) {
-            this.proj = prj;
-            this.ctx = ctx;
-            this.assrts = asserts;
-        }
-
-        @Override
-        public MavenProject project() {
-            return this.proj;
-        }
-
-        @Override
-        public Properties properties() {
-            throw new UnsupportedOperationException("#properties()");
-        }
-
-        @Override
-        public Context context() {
-            return this.ctx;
-        }
-
-        @Override
-        public Properties config() {
-            throw new UnsupportedOperationException("#config()");
-        }
-
-        @Override
-        public MojoExecutor executor() {
-            throw new UnsupportedOperationException("#executor()");
-        }
-
-        @Override
-        public Collection<String> asserts() {
-            return this.assrts;
-        }
-
-        @Override
-        public File basedir() {
-            return this.proj.getBasedir();
-        }
-
-        @Override
-        public File tempdir() {
-            return new File("/tmp");
-        }
-
-        @Override
-        public File outdir() {
-            return this.proj.getBasedir();
-        }
-
-        @Override
-        public Collection<File> testdirs() {
-            return this.proj.getTestCompileSourceRoots().stream()
-                .map(File::new)
-                .toList();
-        }
-
-        @Override
-        public String param(final String name, final String value) {
-            return "";
-        }
-
-        @Override
-        public ClassLoader classloader() {
-            return ClassLoader.getSystemClassLoader();
-        }
-
-        @Override
-        public Collection<String> classpath() {
-            return Collections.emptyList();
-        }
-
-        @Override
-        public Collection<File> files(final String pattern) {
-            return Collections.emptyList();
-        }
-
-        @Override
-        public boolean exclude(final String check, final String name) {
-            return true;
-        }
-
-        @Override
-        public Collection<String> excludes(final String checker) {
-            return Collections.emptyList();
-        }
-
-        @Override
-        public Charset encoding() {
-            return StandardCharsets.UTF_8;
-        }
     }
 }
