+++ src/test/resources/com/qulice/checkstyle/ChecksTest/ProhibitStaticNestedClassesCheck/Invalid-Deep.java
@@ -1,0 +1,11 @@
+/*
+ * This is not a real Java class. It won't be compiled ever. It is used
+ * only as a text resource in integration.ChecksIT.
+ */
+public class InvalidDeep {
+    private final class Inner {
+        private static final class Nested {
+            private int counter;
+        }
+    }
+}
