+++ src/test/resources/com/qulice/checkstyle/ChecksTest/ProhibitStaticNestedClassesCheck/Invalid.java
@@ -1,0 +1,9 @@
+/*
+ * This is not a real Java class. It won't be compiled ever. It is used
+ * only as a text resource in integration.ChecksIT.
+ */
+public class Invalid {
+    private static class Nested {
+        private int counter;
+    }
+}
