+++ src/test/resources/com/qulice/checkstyle/ChecksTest/ProhibitStaticNestedClassesCheck/Valid-SampleTest.java
@@ -1,0 +1,9 @@
+/*
+ * This is not a real Java class. It won't be compiled ever. It is used
+ * only as a text resource in integration.ChecksIT.
+ */
+public class SampleTest {
+    private static final class FakeHelper {
+        private int counter;
+    }
+}
