+++ src/test/resources/com/qulice/checkstyle/ChecksTest/ProhibitStaticNestedClassesCheck/config.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+* SPDX-FileCopyrightText: Copyright (c) 2011-2026 Yegor Bugayenko
+* SPDX-License-Identifier: MIT
+-->
+<!DOCTYPE module PUBLIC "-//Puppy Crawl//DTD Check Configuration 1.2//EN" "http://www.puppycrawl.com/dtds/configuration_1_2.dtd">
+<module name="Checker">
+  <module name="TreeWalker">
+    <module name="com.qulice.checkstyle.ProhibitStaticNestedClassesCheck">
+      <property name="excludeFileNamePattern" value=".*(Test|IT|ITCase)\.java$"/>
+    </module>
+  </module>
+</module>
