+++ src/test/resources/com/qulice/checkstyle/ChecksTest/ProhibitStaticNestedClassesCheck/violations-Deep.txt
@@ -1,0 +1,1 @@
+7:Static class "Nested" must be moved into its own file
