+++ src/test/resources/com/qulice/checkstyle/ChecksTest/ProhibitStaticNestedClassesCheck/violations.txt
@@ -1,0 +1,1 @@
+6:Static class "Nested" must be moved into its own file
